Resolve momotor issue with output incorrect

+++ src/Parser/Parser.java
@@ -43,8 +43,8 @@ public class Parser implements ParserInterface {
         }
 
         try {
-            while (!sc.hasNextFloat()) sc.next();
-            rec.aspectRatio = sc.nextFloat();
+            while (!sc.hasNextDouble()) sc.next();
+            rec.aspectRatio = sc.nextDouble();
         } catch (NoSuchElementException e) {
             throw new IOException("parser.input: no aspect ratio found");
         }
